rename Permission to Permission node with proper interface
